backend/requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
fastapi
uvicorn[standard]
requests
pandas
python-multipart